refactored data class into multiple classes to avoid any circular import issues
