+++ lib/version.py
@@ -1,1 +1,1 @@
-__version_info__ = 'v3.5.1'
+__version_info__ = 'v3.5.0'
