--- ui/data.py
@@ -1,410 +1,0 @@
-import enum
-import sys
-from functools import partial
-
-from PySide2.QtCore import Qt
-from PySide2.QtWidgets import QHeaderView, QPushButton, QComboBox
-
-from lib.config import Config, ConfigType
-from lib.utils import Utils
-from ui.callbcks import Callbacks
-
-
-class Icons(enum.Enum):
-
-    DOWNLOAD_VIDEO = '⬇'
-    PLAY_VIDEO = '▶'
-    OPEN_FOLDER = '⏏'
-    DOWNLOAD_SLIDES = '⬇'
-    DOWNLOAD_CAPTIONS = '㏄'
-    SHOW_SLIDES = '▤'
-    ATTACH_SLIDES = '📎'
-    PAUSE_DOWNLOAD = '❘❘'
-    RESUME_DOWNLOAD = '❘❘▶'
-    VIDEO_PROCESSING = '⧗'
-    VIDEO_DOWNLOADED = '✓'
-    VIDEO_NOT_DOWNLOADED = '⃠'
-    SLIDES_DOWNLOADED = '▤'
-    MOVED_TO = '⇨'
-
-    def __str__(self):
-        return str(self.value)
-
-
-class Labels(enum.Enum):
-    RELOAD = '⟳  Reload'
-    AUTO_ORGANIZE = '⇄  Auto Organize Lectures'
-    COLUMNS = '❘❘❘  Columns'
-    FLIPPED_QUALITY = '☇  Flipped Lecture Quality'
-    QUIT = 'Quit'
-    ACTIONS = 'Actions'
-    COLORSCHEME = 'Color Scheme'
-    VIEW = 'View'
-    VIDEO = 'Video'
-    DOCUMENTATION = 'Documentation...'
-    CHECK_FOR_UPDATES = 'Check for updates...'
-    HELP = 'Help'
-    APPLICATION_TITLE = 'Impartus Downloader'
-    LOGIN_TITLE = 'Login - Impartus'
-
-    def __str__(self):
-        return str(self.value)
-
-
-class Columns:
-    data_columns = {
-        'subjectNameShort': {
-            'alignment': Qt.AlignLeft,
-            'display_name': 'Subject',
-            'editable': True,
-            'hidden': False,
-            'menu_tooltip': 'Subject',
-            'original_values_col': 'subjectName',
-            'resize_policy': QHeaderView.ResizeMode.ResizeToContents,
-            'sortable': True,
-            'title_case': False,
-        },
-        'professorName': {
-            'alignment': Qt.AlignLeft,
-            'display_name': 'Faculty',
-            'editable': False,
-            'hidden': False,
-            'menu_tooltip': 'Faculty Name',
-            'original_values_col': None,
-            'resize_policy': QHeaderView.ResizeMode.ResizeToContents,
-            'sortable': True,
-            'title_case': True,
-        },
-        'topic': {
-            'alignment': Qt.AlignLeft,
-            'display_name': 'Topic',
-            'editable': False,
-            'hidden': False,
-            'menu_tooltip': 'Lecture Topic',
-            'original_values_col': None,
-            'resize_policy': QHeaderView.ResizeMode.Stretch,
-            'sortable': True,
-            'title_case': True,
-        },
-        'seqNo': {
-            'alignment': Qt.AlignRight,
-            'display_name': 'Lecture #',
-            'editable': False,
-            'hidden': False,
-            'menu_tooltip': 'Lecture id',
-            'original_values_col': None,
-            'resize_policy': QHeaderView.ResizeMode.ResizeToContents,
-            'sortable': True,
-            'title_case': False,
-        },
-        'actualDurationReadable': {
-            'alignment': Qt.AlignRight,
-            'display_name': 'Duration',
-            'editable': False,
-            'hidden': False,
-            'menu_tooltip': 'Lecture Duration',
-            'original_values_col': None,
-            'resize_policy': QHeaderView.ResizeMode.ResizeToContents,
-            'sortable': True,
-            'title_case': False,
-        },
-        'tapNToggle': {
-            'alignment': Qt.AlignRight,
-            'display_name': 'Tracks',
-            'editable': False,
-            'hidden': False,
-            'menu_tooltip': 'Number of tracks / view',
-            'original_values_col': None,
-            'resize_policy': QHeaderView.ResizeMode.ResizeToContents,
-            'sortable': True,
-            'title_case': False,
-        },
-        'startDate': {
-            'alignment': Qt.AlignRight,
-            'display_name': 'Date',
-            'editable': False,
-            'hidden': False,
-            'menu_tooltip': 'Lecture Date',
-            'original_values_col': None,
-            'resize_policy': QHeaderView.ResizeMode.ResizeToContents,
-            'sortable': True,
-            'title_case': False,
-        },
-    }
-
-    widget_columns = {
-        'progress_bar': {
-            'alignment': Qt.AlignCenter,
-            'display_name': 'Downloaded?',
-            'editable': False,
-            'hidden': False,
-            'menu_tooltip': 'Download Status',
-            'original_values_col': None,
-            'resize_policy': QHeaderView.ResizeMode.Fixed,
-            'sortable': False,
-            'title_case': False,
-        },
-        'video_actions': {
-            'alignment': Qt.AlignCenter,
-            'display_name': 'Video',
-            'editable': False,
-            'hidden': False,
-            'menu_tooltip': 'Video Actions',
-            'original_values_col': None,
-            'resize_policy': QHeaderView.ResizeMode.ResizeToContents,
-            'sortable': False,
-            'title_case': False,
-        },
-        'slides_actions': {
-            'alignment': Qt.AlignRight,
-            'display_name': 'Slides',
-            'editable': False,
-            'hidden': False,
-            'menu_tooltip': 'Slides Actions',
-            'original_values_col': None,
-            'resize_policy': QHeaderView.ResizeMode.ResizeToContents,
-            'sortable': False,
-            'title_case': False,
-        },
-    }
-
-    hidden_columns = {
-        'ttid': {
-            'alignment': Qt.AlignRight,
-            'display_name': 'ttid',
-            'editable': False,
-            'hidden': True,
-            'menu_tooltip': 'ttid',
-            'original_values_col': None,
-            'resize_policy': QHeaderView.ResizeMode.ResizeToContents,
-            'sortable': True,
-            'title_case': False,
-        },
-    }
-
-    @classmethod
-    def get_display_columns(cls):
-        return [*Columns.data_columns, *Columns.widget_columns]
-
-    @classmethod
-    def get_column_index_by_key(cls, key_name):
-        return [*Columns.data_columns.keys(), *Columns.widget_columns.keys(), *Columns.hidden_columns.keys()].index(
-            key_name) + 1
-
-    @classmethod
-    def get_column_index_by_display_name(cls, display_name):
-        for index, value in enumerate(
-                [*Columns.data_columns.values(), *Columns.widget_columns.values(), *Columns.hidden_columns.values()]):
-            if value['display_name'] == display_name:
-                return index + 1
-
-    @classmethod
-    def get_columns_count(cls):
-        return 1 + len(Columns.data_columns) + len(Columns.widget_columns) + len(Columns.hidden_columns)
-
-    @classmethod
-    def get_button_order(cls):
-        return 1 + len(Columns.data_columns) + len(Columns.widget_columns) + len(Columns.hidden_columns)
-
-
-class ConfigKeys(enum.Enum):
-    URL = 'impartus_url'
-    EMAIL = 'login_email'
-    PASSWORD = 'password'
-    TARGET_DIR = 'target_dir'
-    CONFIG_DIR = 'config_dir'
-    ALLOWED_EXT = 'allowed_ext'
-    VIDEO_PATH = 'video_path'
-    SLIDES_PATH = 'slides_path'
-    CAPTIONS_PATH = 'captions_path'
-    COLORSCHEME_DEFAULT = 'default'
-    RESIZE_POLICY = 'resize_policy'
-
-    def __str__(self):
-        return str(self.value)
-
-
-class IconFiles(enum.Enum):
-
-    SORT_UP_ARROW = 'images/sort-up.png'
-    SORT_DOWN_ARROW = 'images/sort-down.png'
-    EDITABLE_BLUE = 'images/editable-blue.png'
-    EDITABLE_RED = 'images/editable-red.png'
-    APP_LOGO = 'images/logo.png'
-
-    def __str__(self):
-        return str(self.value)
-
-
-class DocFiles(enum.Enum):
-
-    HELPDOC = 'docs/helpdoc.pdf'
-
-    def __str__(self):
-        return str(self.value)
-
-
-class ActionItems:
-    video_actions = {
-        'download_video': {
-            'tooltip': '{} Download Video'.format(Icons.DOWNLOAD_VIDEO.value),
-            'text': Icons.DOWNLOAD_VIDEO.value,
-            'type': QPushButton,
-        },
-        'play_video': {
-            'tooltip': '{} Play Video'.format(Icons.PLAY_VIDEO.value),
-            'text': Icons.PLAY_VIDEO.value,
-            'type': QPushButton,
-        },
-        'download_chats': {
-            'tooltip': '{} Download Lecture Chats'.format(Icons.DOWNLOAD_CAPTIONS.value),
-            'text': Icons.DOWNLOAD_CAPTIONS.value,
-            'type': QPushButton,
-        },
-    }
-    slides_actions = {
-        'show_slides': {
-            'tooltip': '{} Show Backpack Slides'.format(Icons.SHOW_SLIDES.value),
-            'text': Icons.DOWNLOAD_SLIDES.value,
-            'type': QComboBox,
-        },
-        'download_slides': {
-            'tooltip': '{} Download Backpack Slides'.format(Icons.DOWNLOAD_SLIDES.value),
-            'text': Icons.DOWNLOAD_SLIDES.value,
-            'type': QPushButton,
-        },
-        'open_folder': {
-            'tooltip': '{} Open Folder'.format(Icons.OPEN_FOLDER.value),
-            'text': Icons.OPEN_FOLDER.value,
-            'type': QPushButton,
-        },
-        'attach_slides': {
-            'tooltip': '{} Attach Slides'.format(Icons.ATTACH_SLIDES.value),
-            'text': Icons.ATTACH_SLIDES.value,
-            'type': QPushButton,
-        },
-    }
-
-    @classmethod
-    def get_action_item_index(cls, action_item, field):    # noqa
-        for i, key in enumerate(getattr(ActionItems, action_item).keys()):
-            if key == field:
-                return i
-
-
-class MenuItems:
-
-    conf = Config.load(ConfigType.IMPARTUS)
-    menu_items = {
-        'Actions': {
-            'Login': {
-                'shortcut': 'Ctrl+L',
-                'status_tip': 'Login to Impartus',
-                'callback': Callbacks().on_login_click,
-            },
-            'Reload': {
-                'shortcut': 'Ctrl+R',
-                'status_tip': 'Reload Table Content',
-                'callback': Callbacks().on_reload_click,
-            },
-            'Auto Organize': {
-                'shortcut': 'Ctrl+/',
-                'status_tip': 'Rename lecture videos, download missing captions...',
-                'callback': Callbacks().on_auto_organize_click,
-            },
-            'sep': {
-                'type': 'separator',
-            },
-            'Quit': {
-                'shortcut': 'Ctrl+Q',
-                'status_tip': 'Quit Application',
-                'callback': partial(sys.exit, 0)
-            },
-        },
-        'View': {
-            'Columns': {
-                'shortcut': None,
-                'status_tip': None,
-                'type': 'list',
-                'status': 'disabled',
-                'behavior': 'checkall',
-                'child_items': [x['display_name']
-                                for x in [*Columns.data_columns.values(), *Columns.widget_columns.values()]],
-                'child_callbacks': [partial(Callbacks().on_column_click, key)
-                                    for key in [*Columns.data_columns.keys(), *Columns.widget_columns.keys()]]
-            },
-            'sep': {
-                'type': 'separator',
-            },
-            'Search': {
-                'shortcut': 'Ctrl+F',
-                'status_tip': 'Search Content...',
-                'callback': Callbacks().on_search_click,
-            },
-        },
-        'Video': {
-            'Flipped Video Lecture Quality': {
-                'shortcut': None,
-                'type': 'list',
-                'status_tip': None,
-                'status': 'disabled',
-                'behavior': 'singleselect',
-                'child_items': [x for x in conf.get('flipped_lecture_quality_order')],
-                'default': conf.get('video_quality'),
-                'child_callbacks': [partial(Callbacks().on_video_quality_click, video_quality)
-                                    for video_quality in conf.get('flipped_lecture_quality')]
-            },
-            'sep': {
-                'type': 'separator',
-            },
-            'Download Video': {
-                'shortcut': 'Ctrl+J',
-                'status_tip': 'Download Lecture Video',
-                'callback': Callbacks().on_download_video_click,
-            },
-            'Play Video': {
-                'shortcut': 'Ctrl+P',
-                'status_tip': 'Play Lecture Video',
-                'callback': Callbacks().on_play_video_click,
-            },
-            'Download Lecture Chats': {
-                'shortcut': 'Shift+Ctrl+J',
-                'status_tip': 'Download Lecture Chats',
-                'callback': Callbacks().on_download_chats_click,
-            },
-        },
-        'Slides': {
-            'Download Backpack Slides': {
-                'shortcut': 'Ctrl+K',
-                'status_tip': 'Download Lecture Slides',
-                'callback': Callbacks().on_download_slides_click,
-            },
-            'Open Folder': {
-                'shortcut': 'Ctrl+O',
-                'status_tip': 'Open Lecture Content Folder',
-                'callback': Callbacks().on_open_folder_click,
-            },
-            'Attach Lecture Slides': {
-                'shortcut': 'Shift+Ctrl+O',
-                'status_tip': 'Attach Downloaded Slides to a Lecture',
-                'callback': Callbacks().on_attach_slides_click,
-            },
-        },
-        'Help': {
-            'Help': {
-                'shortcut': None,
-                'status_tip': 'Help docs',
-                'callback': partial(Utils.open_file, 'docs/helpdoc.pdf'),
-            },
-            'Check For Updates': {
-                'shortcut': None,
-                'status_tip': 'Check for new version, release notes...',
-                'callback': Callbacks().on_check_for_updates_click,
-            },
-        }
-    }
-
-
-class SearchDirection(enum.Enum):
-    FORWARD = 1
-    BACKWARD = -1
