+++ ui/data/configkeys.py
@@ -1,0 +1,19 @@
+import enum
+
+
+class ConfigKeys(enum.Enum):
+    URL = 'impartus_url'
+    EMAIL = 'login_email'
+    PASSWORD = 'password'
+    TARGET_DIR = 'target_dir'
+    CONFIG_DIR = 'config_dir'
+    ALLOWED_EXT = 'allowed_ext'
+    VIDEO_PATH = 'video_path'
+    SLIDES_PATH = 'slides_path'
+    CAPTIONS_PATH = 'captions_path'
+    COLORSCHEME_DEFAULT = 'default'
+    RESIZE_POLICY = 'resize_policy'
+
+    def __str__(self):
+        return str(self.value)
+
