+++ ui/data/docs.py
@@ -1,0 +1,10 @@
+import enum
+
+
+class Docs(enum.Enum):
+
+    HELPDOC = 'docs/helpdoc.pdf'
+
+    def __str__(self):
+        return str(self.value)
+
