+++ ui/data/searchdirections.py
@@ -1,0 +1,6 @@
+import enum
+
+
+class SearchDirection(enum.Enum):
+    FORWARD = 1
+    BACKWARD = -1
