+++ ui/search.py
@@ -1,7 +1,7 @@
 from PySide2 import QtCore
 from PySide2.QtWidgets import QHBoxLayout, QWidget, QLabel, QLineEdit, QMainWindow, QTableWidget
 
-from ui.data import SearchDirection
+from ui.data.searchdirections import SearchDirection
 
 
 class SearchBox:
